error case is supported

+++ src/ft_printf/ft_printf_utils.c
@@ -6,7 +6,7 @@
 /*   By: shogura <shogura@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/28 17:31:25 by shogura           #+#    #+#             */
-/*   Updated: 2022/05/12 16:51:12 by shogura          ###   ########.fr       */
+/*   Updated: 2022/05/13 13:43:24 by shogura          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ size_t	ft_strlen(const char *s)
 	return (len);
 }
 
-int	get_digits(long num, int base)
+int	get_digits(long long num, int base)
 {
 	int	digits;
